Rename fcoin/endpoint => fcoin/api (Connection & Request)
