+++ fcoin/api/request.go
@@ -1,4 +1,4 @@
-package endpoint
+package api
 
 import (
 	"fmt"
